# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QR-API",
    "version": "1.0.0",
    "description": "API de WhatsApp via QR Code da Flipt. Conecta qualquer numero escaneando um QR uma vez e permite enviar mensagens ilimitadas por assinatura fixa. Inclui anti-ban automatico, webhooks com HMAC, idempotencia de envio, erros estruturados e sandbox gratuita.",
    "contact": {
      "name": "Suporte Flipt",
      "url": "https://flipt.com.br/suporte",
      "email": "suporte@flipt.com.br"
    },
    "license": {
      "name": "Proprietario"
    }
  },
  "servers": [
    {
      "url": "https://api.qrapi.flipt.com.br/v1",
      "description": "Standalone (produto qr-api publico)"
    },
    {
      "url": "https://api.flipt.com.br/api/qrapi/v1",
      "description": "Embedded (dentro do Followfy/Flipt)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Chave de API no formato sk_test_xxx (sandbox) ou sk_live_xxx (producao). Nunca envie na URL."
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["type", "code", "message"],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "authentication_error",
                  "permission_error",
                  "invalid_request_error",
                  "not_found_error",
                  "idempotency_error",
                  "rate_limit_error",
                  "instance_error",
                  "api_error"
                ],
                "description": "Categoria de erro para tratamento programatico."
              },
              "code": {
                "type": "string",
                "description": "Codigo especifico do erro.",
                "example": "missing_parameter"
              },
              "message": {
                "type": "string",
                "description": "Descricao legivel para exibicao ou log.",
                "example": "O campo 'to' e obrigatorio."
              },
              "param": {
                "type": "string",
                "description": "Campo invalido/ausente (presente em invalid_request_error).",
                "example": "to"
              },
              "doc_url": {
                "type": "string",
                "format": "uri",
                "description": "Link direto para a pagina de documentacao relevante.",
                "example": "https://docs.qrapi.flipt.com.br/send-messages"
              }
            }
          },
          "request_id": {
            "type": "string",
            "description": "Identificador unico desta requisicao. Inclua ao contatar o suporte.",
            "example": "req_8f2a1c"
          }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "key_abc" },
          "name": { "type": "string", "example": "Integracao ERP" },
          "prefix": {
            "type": "string",
            "description": "Primeiros caracteres da chave para identificacao visual.",
            "example": "sk_live_4eC3"
          },
          "scopes": { "type": "array", "items": { "type": "string" } },
          "lastUsedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "required": ["name", "scopes"],
        "properties": {
          "name": { "type": "string" },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["messages:send", "messages:read", "instances:read", "instances:write", "webhooks:write", "phones:read", "*"]
            }
          },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "PhoneExistsResponse": {
        "type": "object",
        "properties": {
          "number": { "type": "string", "example": "5511999998888" },
          "exists": { "type": "boolean" },
          "jid": { "type": "string", "description": "JID do WhatsApp se exists=true.", "example": "5511999998888@s.whatsapp.net" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Autenticacao invalida.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Scope insuficiente.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Recurso nao encontrado.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  },
  "paths": {
...
    "/phones/{number}/exists": {
      "parameters": [
        {
          "name": "number",
          "in": "path",
          "required": true,
          "schema": { "type": "string" },
          "description": "Numero em E.164 sem + (ex.: 5511999998888).",
          "example": "5511999998888"
        }
      ],
      "get": {
        "operationId": "checkPhone",
        "summary": "Verificar numero",
        "description": "Verifica se um numero tem conta ativa no WhatsApp.",
        "tags": ["Numeros"],
        "responses": {
          "200": { "description": "Resultado da verificacao.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PhoneExistsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/phones/exists": {
      "post": {
        "operationId": "checkPhonesBulk",
        "summary": "Verificar numeros em lote",
        "tags": ["Numeros"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["numbers"],
                "properties": {
                  "numbers": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["5511999998888", "5511999990001"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultados em lote.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "array", "items": { "$ref": "#/components/schemas/PhoneExistsResponse" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/contacts/{number}": {
      "parameters": [
        { "name": "number", "in": "path", "required": true, "schema": { "type": "string" }, "example": "5511999998888" }
      ],
      "get": {
        "operationId": "getContact",
        "summary": "Perfil do contato",
        "description": "Retorna nome publico e URL da foto de perfil do contato.",
        "tags": ["Numeros"],
        "responses": {
          "200": {
            "description": "Perfil do contato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "number": { "type": "string" },
                    "name": { "type": "string" },
                    "profilePictureUrl": { "type": "string", "format": "uri" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/keys": {
      "post": {
        "operationId": "createApiKey",
        "summary": "Cria API key",
        "description": "Cria uma nova chave de API. A chave completa e retornada apenas nesta resposta.",
        "tags": ["Chaves"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateApiKeyRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Chave criada. Guarde o campo key: nao sera exibido novamente.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiKey" },
                    { "type": "object", "properties": { "key": { "type": "string", "description": "Chave completa (so aparece aqui).", "example": "sk_live_4eC39HqLyjWDarjtT1zdp7dc" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "get": {
        "operationId": "listApiKeys",
        "summary": "Lista chaves",
        "description": "Lista chaves com prefixo e lastUsedAt. Nunca retorna a chave completa.",
        "tags": ["Chaves"],
        "responses": {
          "200": {
            "description": "Lista de chaves.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/keys/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "delete": {
        "operationId": "revokeApiKey",
        "summary": "Revoga chave",
        "description": "Revogacao imediata. Requisicoes em andamento sao rejeitadas na proxima validacao.",
        "tags": ["Chaves"],
        "responses": {
          "204": { "description": "Chave revogada." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
...
}